pyproject.toml
[tool.poetry]
name = "umbria_one"
version = "0.0.1"
description = "Arbitrage strategy engine"
authors = [
  "edazizovv",
]
license = "MIT"
readme = "README.md"
packages = [{include = "umbria_one"}]

[tool.poetry.dependencies]
python = "^3.14"
xlsxwriter = "^3.2.9"
openpyxl = "^3.1.5"
numpy = "^2.4.2"
pandas = "^3.0.1"
scipy = "^1.17.0"
matplotlib = "^3.10.8"
seaborn = "^0.13.2"
jupyter = "^1.1.1"
python-dotenv = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"